Handle self-closing subfield tag
If an empty (self-closing) tag is received, an error is currently
thrown as it cannot parse the `None` value that lxml produces.

This update checks for a none value returned in the Subfield `__init__`
and replaces it with an empty string if found.

It's unclear to me if a field with a single empty subfield really should
be in a MARCXML record, but that is (unfortunately) not something
that can be controlled for.

This commit also includes a new test with a new XML file that includes
a `520` field that contains this case